creación de la entidad de Orden

+++ BichoPedia/src/main/java/com/salesianostriana/dam/bichopedia/model/Usuario.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 public class Usuario {
 	@Id
 	@GeneratedValue
-	private int id_usuario;
+	private Long id_usuario;
 	
 	private String nombre;
 	private String apellidos;
